--- akka-sample-cluster-istio-scala/frontend/src/main/resources/application.conf
@@ -1,25 +1,0 @@
-akka {
-  actor {
-    provider = cluster
-  }
-
-  remote {
-    log-remote-lifecycle-events = off
-  }
-
-  cluster {
-    # auto downing is NOT safe for production deployments.
-    # you may want to use it during development, read more about it in the docs.
-    auto-down-unreachable-after = 10s
-  }
-}
-
-# Disable legacy metrics in akka-cluster.
-akka.cluster.metrics.enabled=off
-
-# Enable metrics extension in akka-cluster-metrics.
-akka.extensions=["akka.cluster.metrics.ClusterMetricsExtension"]
-
-# Sigar native library extract location during tests.
-# Note: use per-jvm-instance folder when running multiple jvm on one host.
-akka.cluster.metrics.native-library-extract-folder=${user.dir}/target/native
