--- akka-sample-cluster-istio-scala/frontend/src/main/resources/factorial.conf
@@ -1,27 +1,0 @@
-include "application"
-
-akka.cluster.min-nr-of-members = 3
-
-akka.cluster.role {
-  frontend.min-nr-of-members = 1
-  backend.min-nr-of-members = 2
-}
-
-akka.actor.deployment {
-  /factorialFrontend/factorialBackendRouter = {
-    # Router type provided by metrics extension. 
-    router = cluster-metrics-adaptive-group
-    # Router parameter specific for metrics extension.
-    # metrics-selector = heap
-    # metrics-selector = load
-    # metrics-selector = cpu
-    metrics-selector = mix
-    #
-    routees.paths = ["/user/factorialBackend"]
-    cluster {
-      enabled = on
-      use-role = backend
-      allow-local-routees = off
-    }
-  }
-}
