--- akka-sample-cluster-istio-scala/frontend/src/main/scala/sample/cluster/factorial/frontend/FactorialFrontend.scala
@@ -1,82 +1,0 @@
-package sample.cluster.factorial.frontend
-
-import scala.concurrent.duration._
-import com.typesafe.config.ConfigFactory
-import akka.actor.Actor
-import akka.actor.ActorLogging
-import akka.actor.ActorSystem
-import akka.actor.Props
-import akka.cluster.Cluster
-import akka.routing.FromConfig
-import akka.actor.ReceiveTimeout
-import scala.util.Try
-import scala.concurrent.Await
-
-class FactorialFrontend(upToN: Int, repeat: Boolean) extends Actor with ActorLogging {
-
-  val backend = context.actorOf(FromConfig.props(),
-    name = "factorialBackendRouter")
-
-  override def preStart(): Unit = {
-    sendJobs()
-    if (repeat) {
-      context.setReceiveTimeout(10.seconds)
-    }
-  }
-
-  def receive = {
-    case (n: Int, factorial: BigInt) =>
-      if (n == upToN) {
-        log.debug("{}! = {}", n, factorial)
-        if (repeat) sendJobs()
-        else context.stop(self)
-      }
-    case ReceiveTimeout =>
-      log.info("Timeout")
-      sendJobs()
-  }
-
-  def sendJobs(): Unit = {
-    log.info("Starting batch of factorials up to [{}]", upToN)
-    1 to upToN foreach { backend ! _ }
-  }
-}
-
-object FactorialFrontend {
-  def main(args: Array[String]): Unit = {
-    val upToN = 200
-
-    val config = ConfigFactory.parseString("akka.cluster.roles = [frontend]").
-      withFallback(ConfigFactory.load("factorial"))
-
-    val systemName = sys.props.get("sample.cluster.actor-system-name").getOrElse(
-      throw new IllegalArgumentException("Property sample.cluster.actor-system-name must be defined")
-    )
-
-    val system = ActorSystem(systemName, config)
-    system.log.info("Factorials will start when 2 backend members in the cluster.")
-
-    Cluster(system) registerOnMemberUp {
-      system.actorOf(Props(classOf[FactorialFrontend], upToN, true),
-        name = "factorialFrontend")
-    }
-
-    Cluster(system).registerOnMemberRemoved {
-      // exit JVM when ActorSystem has been terminated
-      system.registerOnTermination(System.exit(0))
-      // shut down ActorSystem
-      system.terminate()
-
-      // In case ActorSystem shutdown takes longer than 10 seconds,
-      // exit the JVM forcefully anyway.
-      // We must spawn a separate thread to not block current thread,
-      // since that would have blocked the shutdown of the ActorSystem.
-      new Thread {
-        override def run(): Unit = {
-          if (Try(Await.ready(system.whenTerminated, 10.seconds)).isFailure)
-            System.exit(-1)
-        }
-      }.start()
-    }
-  }
-}
