+++ akka-sample-cluster-istio-scala/src/main/resources/application.conf
@@ -1,0 +1,48 @@
+akka {
+  extensions=["akka.cluster.metrics.ClusterMetricsExtension"]
+
+  metrics.native-library-extract-folder=${user.dir}/target/native
+
+  actor {
+    provider = cluster
+
+    deployment {
+      /factorial/frontend/backendRouter = {
+        # Router type provided by metrics extension.
+        router = cluster-metrics-adaptive-group
+        # Router parameter specific for metrics extension.
+        # metrics-selector = heap
+        # metrics-selector = load
+        # metrics-selector = cpu
+        metrics-selector = mix
+        #
+        routees.paths = ["/user/factorial/backend"]
+        cluster {
+          enabled = on
+          use-role = backend
+          allow-local-routees = off
+        }
+      }
+    }
+  }
+
+  remote {
+    log-remote-lifecycle-events = off
+  }
+
+  cluster {
+    # auto downing is NOT safe for production deployments.
+    # you may want to use it during development, read more about it in the docs.
+    auto-down-unreachable-after = 10s
+
+    # Disable legacy metrics in akka-cluster.
+    metrics.enabled=off
+
+    min-nr-of-members = 3
+
+    role {
+      frontend.min-nr-of-members = 1
+      backend.min-nr-of-members = 2
+    }
+  }
+}
